memops: only create phi-sources when it's sure they're needed
Currently, when searching for the dominating parents, phi-sources are
optimistically created as soon as the parent memop is found. Then, if
all these parents have been found, the corresponding phi-node is created,
otherwise the phi-sources need to be deleted.

This is annoying because it conceptually complicates things and very
often it's wasted work because it's frequent that no dominating parent
can be found.

So, change to only keeping track of the parent memops and create
the corresponding phi-sources together with the phi-node, when all
parents have been found.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>