memops: change link_phi() prototype
link_phi() is a small helper used to set the informations linking
a phi-source and its phi-node. The phi-source is passed as the
pseudo but the corresponding is also needed.

Since, it's more flexible to dereference an instruction to get its
target pseudo than doing the opposite, change the prototype of the
helper to use the phi-source instruction instead of the pseudo.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>