memops: do parent simplification outside the list walking
There is two kind of load simplifications:
1) with a dominating memop in the same BB (local)
2) with dominating memops in the parent BBs.
While the local ones can be done by simply rewriting the load,
in general, the non-local ones require the insertion of a phi-node
(which should be created at the top of the BB).

But this insertion conflict with he list walking done to find the
loads to be simplified.

So, fix this by using an intermediate worklist:
* if a no local dominating memop is found, store a pointer to the
  load instruction and process the loads of this list only afer
  the current BB have been scanned for loads.

Note: currently phi-nodes are not (yet) inserted at the top of the
      BB and this patch is one of the step needed to fix that.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>