llvm: do not duplicate strings and use their length in struct string
In 2 places, we duplicate the storage for a string (with strdup) and we
also calculate its length via strlen(). Both operation are unneeded
as the length is already calculated in the struct string and the pointer
to the string data can be safely reused since Sparse will not modify or free it.

Signed-off-by: Luc Van Oostenryck <lucvoo@kernel.org>